% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_acknowledgements.R
\name{acknowledgements_module_server}
\alias{acknowledgements_module_server}
\title{Title}
\usage{
acknowledgements_module_server(id)
}
\arguments{
\item{id}{identification}
}
\description{
Title
}
