% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_acknowledgements.R
\name{acknowledgements_module_ui}
\alias{acknowledgements_module_ui}
\title{UI module acknowledgements}
\usage{
acknowledgements_module_ui(id)
}
\arguments{
\item{id}{module id}
}
\description{
UI module acknowledgements
}
