% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_data.R
\docType{data}
\name{ca_spain_gj}
\alias{ca_spain_gj}
\title{ca_spain_gj data}
\format{
An object of class \code{SpatialPolygonsDataFrame} with 20 rows and 9 columns.
}
\usage{
ca_spain_gj
}
\description{
ca_spain_gj data
}
\keyword{datasets}
