% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_data.R
\docType{data}
\name{catalunya_map}
\alias{catalunya_map}
\title{catalunya_map data}
\format{
An object of class \code{SpatialPolygonsDataFrame} with 50 rows and 3 columns.
}
\usage{
catalunya_map
}
\description{
catalunya_map data
}
\keyword{datasets}
