% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{check_mutations}
\alias{check_mutations}
\title{Title}
\usage{
check_mutations(mut, all_pos)
}
\arguments{
\item{mut}{mutation columns}

\item{all_pos}{all positions}
}
\value{
list
}
\description{
Title
}
