% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{efforts_by_center}
\alias{efforts_by_center}
\title{Plots sequencing efforts by center}
\usage{
efforts_by_center(df, pos = "stack")
}
\arguments{
\item{pos}{}
}
\value{
plotly
}
\description{
Plots sequencing efforts by center
}
