% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map_data}
\alias{map_data}
\title{Map data}
\usage{
map_data(df, my_map_data, ca_inhabitants)
}
\arguments{
\item{df}{tibble}

\item{my_map_data}{map geojson}

\item{ca_inhabitants}{inhabitant values by C.A.}
}
\value{
list
}
\description{
Map data
}
