% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_overview_2.R
\name{overview2_module_server}
\alias{overview2_module_server}
\title{Sever module overview}
\usage{
overview2_module_server(id)
}
\arguments{
\item{id}{module id}
}
\description{
Sever module overview
}
