% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_mutation_line}
\alias{plot_mutation_line}
\title{Title}
\usage{
plot_mutation_line(inp_list, region = "Spain", mut_pos = "S:1251", mut = "V")
}
\arguments{
\item{var_col}{}
}
\value{

}
\description{
Title
}
