% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_mutations}
\alias{plot_mutations}
\title{Title}
\usage{
plot_mutations(
  df,
  mut_pos = "ORF1b:P314",
  var = "counts",
  pal = "mg",
  plotly = TRUE
)
}
\arguments{
\item{mut_pos}{}
}
\value{

}
\description{
Title
}
