% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_mutations_2}
\alias{plot_mutations_2}
\title{Title}
\usage{
plot_mutations_2(
  inp_list,
  region = "Spain",
  mut_pos = "S:1251",
  var = "counts",
  pal = "mg",
  plotly = TRUE
)
}
\arguments{
\item{region}{C.A or Spain}

\item{mut_pos}{position to plot}

\item{var}{counts or freq}

\item{pal}{mg or RColorBrewer palette}

\item{plotly}{TRUE or FALSE}

\item{df}{df}
}
\value{
plot
}
\description{
Title
}
