% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_vairants}
\alias{plot_vairants}
\title{Plot variants}
\usage{
plot_vairants(
  df,
  type = "density",
  var = "freq",
  pal = "mg",
  pal_dir = -1,
  clade = NULL,
  plotly = TRUE
)
}
\arguments{
\item{df}{output of function prepro_variants}

\item{type}{Class of plot. Options: density or bar}

\item{var}{variable to plot. Options: "freq" and "counts"}

\item{pal}{Used palette}

\item{pal_dir}{Palete direction order. Options: 1 and -1}

\item{clade}{List with clade to plot}

\item{plotly}{boolean indicating if output plot must be plotly}
}
\value{
plotly or ggplot object
}
\description{
Plot variants
}
