% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_variant_by_com}
\alias{plot_variant_by_com}
\title{Plot variant by C.A.}
\usage{
plot_variant_by_com(df, variant, var_col)
}
\arguments{
\item{df}{tibble}

\item{variant}{NCClade to plot}
}
\value{
plotly
}
\description{
Plot variant by C.A.
}
