% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{prepro_mutations}
\alias{prepro_mutations}
\title{Preprocessing mutations}
\usage{
prepro_mutations(df, ca = "Spain")
}
\arguments{
\item{df}{df}

\item{ca}{CA or spain}
}
\value{
df
}
\description{
Preprocessing mutations
}
