% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{prepro_variants}
\alias{prepro_variants}
\title{Pre-processing variants data}
\usage{
prepro_variants(df, ca = "Spain", var_anno = "NCClade")
}
\arguments{
\item{df}{tibble}

\item{ca}{all or autonomous community}

\item{var_anno}{variant annotation column. ex. "NCClade" or "pangolin_lineage"}
}
\value{
tibble
}
\description{
Pre-processing variants data
}
