% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_acknowledgements.R
\name{acknowledgements_module_server}
\alias{acknowledgements_module_server}
\title{Server acknowledgements module}
\usage{
acknowledgements_module_server(id)
}
\arguments{
\item{id}{module id}
}
\description{
Server acknowledgements module
}
