% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{option_mutation}
\alias{option_mutation}
\title{Option mutations}
\usage{
option_mutation(inp_list, gene_pos)
}
\arguments{
\item{inp_list}{list}

\item{gene_pos}{gene:pos}
}
\value{
character vector
}
\description{
Option mutations
}
