% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_mutation_line}
\alias{plot_mutation_line}
\title{Plot mutation line}
\usage{
plot_mutation_line(inp_list, region = "Spain", mut_pos = "S:1251", mut = "V")
}
\arguments{
\item{inp_list}{mutation list}

\item{region}{region}

\item{mut_pos}{muatatio pos}

\item{mut}{variant}
}
\value{
plotly
}
\description{
Plot mutation line
}
