% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_mutations}
\alias{plot_mutations}
\title{Title}
\usage{
plot_mutations(
  df,
  mut_pos = "ORF1b:P314",
  var = "counts",
  pal = "mg",
  plotly = TRUE
)
}
\arguments{
\item{df}{df}

\item{mut_pos}{position to plot}

\item{var}{counts or freq}

\item{pal}{mg or RColorBrewer palette}

\item{plotly}{TRUE or FALSE}
}
\value{
plot
}
\description{
Title
}
